Add files via upload
these are the files related to Parkinson's diseases detection using machine learning , here I used Streamlit to deploy as a web application. this project is done using the XGBoost algorithm. make sure that proper libraries are installed to run this code and that the text file pdd. grph is used to generate the graphs try to run the code in google collab and Anaconda